+++ SivaEftCheat/SivaEftCheat/Options/MiscOptions.cs
@@ -31,7 +31,7 @@ namespace SivaEftCheat.Options
         public static bool InstantHit = false;
         public static bool DontMoveWeaponCloser = true;
 
-        public static float SpeedHackValue = 0f;
+        public static double SpeedHackValue = 0f;
         public static bool DrawHud = true;
     }
 }
